Refactor kafka_message_array
- Delete kafka_message_list, unnecesary copies
- Simplified zz_http2k parsing. No more alloc/copies than needed
- rdkafka_produce errors last timestamp now saved in a coherent way
- use new rdkafka message delivered callback, the previous is deprecated

+++ tests/zz_http2k_tests.c
@@ -215,7 +215,7 @@ static void n2k_consume_kafka_msgs(rd_kafka_t *rk,
 			assert_return_code(size, errno);                       \
 			size++;                                                \
 			if (NULL == ret) {                                     \
-				ret = alloca(size);                            \
+				ret = alloca((size_t)size);                    \
 			}                                                      \
 		}                                                              \
 		ret;                                                           \
